found offby1error

+++ Application/SamplingTypes.h
@@ -1,1 +1,1 @@
-typedef float sample_t ;
+typedef double sample_t ;
